backend/requirements.txt
# FastAPI Medical App Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Payment Processing
stripe==7.8.0

# HTTP & Networking
httpx==0.25.2
aiofiles==23.2.1

# Data Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# Date & Time
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Production Server
gunicorn==21.2.0

# Optional: Monitoring & Logging
structlog==23.2.0

# Optional: Background Tasks
celery==5.3.4
redis==5.0.1